Update 1. Two Sum.cpp
Very Detailed Answer: Finding Two Numbers in an Array that Add Up to a Target {Header1}
Question Summary
Given an array of integers nums and an integer target, the task is to find and return the indices of two numbers in the array that add up to the given target. It is guaranteed that there is exactly one solution, and each element in the array can only be used once.

Answer
To solve this problem efficiently, we can utilize a hash map to store the elements of the array along with their indices. By iterating through the array and checking if the complement of the current element (i.e., target - nums[i]) exists in the hash map, we can find the pair of indices that sum up to the target.